så bookingactivity får booking id med

+++ src/main/java/bigbowl/bookingactivity/BookingActivityRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface BookingActivityRepository extends JpaRepository<BookingActivity, Integer> {
+public interface BookingActivityRepository extends JpaRepository<BookingActivity, Long> {
     List<BookingActivity> findByStartTimeBetweenAndEndTimeBetween(LocalDateTime startTime, LocalDateTime endTime, LocalDateTime startTime1, LocalDateTime endTime1);
 }
